engageMessage edit bug fixed
