--- src/modules/engage/components/FormBase.js
@@ -1,33 +1,0 @@
-import { Component } from 'react';
-import PropTypes from 'prop-types';
-
-const propTypes = {
-  brands: PropTypes.array,
-  users: PropTypes.array,
-  save: PropTypes.func,
-  kind: PropTypes.string,
-  message: PropTypes.object
-};
-
-class FormBase extends Component {
-  constructor(props) {
-    super(props);
-
-    this.save = this.save.bind(this);
-  }
-
-  save(type, e) {
-    const doc = this.generateDoc(e);
-    if (type === 'live') {
-      this.props.save({ isLive: true, isDraft: false, ...doc });
-    } else if (type === 'draft ') {
-      this.props.save({ isLive: false, isDraft: true, ...doc });
-    } else {
-      this.props.save(doc);
-    }
-  }
-}
-
-FormBase.propTypes = propTypes;
-
-export default FormBase;
