+++ src/modules/engage/components/Segments.js
@@ -16,16 +16,10 @@ class Segments extends Component {
     super(props);
 
     this.state = {
-      chosenSegment: ''
+      chosenSegment: props.defaultValue
     };
 
     this.onClickSegment = this.onClickSegment.bind(this);
-  }
-
-  componentDidMount() {
-    if (this.props.defaultValue !== '') {
-      this.setState({ chosenSegment: this.props.defaultValue });
-    }
   }
 
   onClickSegment(segmentId) {
