+++ src/modules/engage/containers/AutoAndManualForm.js
@@ -5,7 +5,7 @@ import { compose, graphql } from 'react-apollo';
 import gql from 'graphql-tag';
 import { AutoAndManualForm } from '../components';
 import { queries, mutations } from '../graphql';
-import withFormMutations from './withFormMutations';
+import { withRouter } from 'react-router';
 
 const AutoAndManualFormContainer = props => {
   const {
@@ -17,11 +17,48 @@ const AutoAndManualFormContainer = props => {
     customerCountsQuery,
     engageMessageDetailQuery,
     brandsQuery,
-    kind
+    kind,
+    messageId,
+    usersQuery,
+    history,
+    addMutation,
+    editMutation
   } = props;
+
+  if (
+    engageMessageDetailQuery.loading ||
+    brandsQuery.loading ||
+    usersQuery.loading
+  ) {
+    return false;
+  }
 
   const message = engageMessageDetailQuery.engageMessageDetail || {};
   const brands = brandsQuery.brands || [];
+  const users = usersQuery.users || [];
+
+  const doMutation = (mutation, variables) => {
+    mutation({
+      variables
+    })
+      .then(() => {
+        Alert.success('Congrats');
+        history.push('/engage');
+      })
+      .catch(error => {
+        Alert.error(error.message);
+      });
+  };
+
+  const save = doc => {
+    doc.kind = message.kind ? message.kind : kind;
+
+    if (messageId) {
+      return doMutation(editMutation, { ...doc, _id: messageId });
+    }
+
+    return doMutation(addMutation, doc);
+  };
 
   const customerCounts = customerCountsQuery.customerCounts || {
     all: 0,
@@ -65,13 +102,18 @@ const AutoAndManualFormContainer = props => {
     customerCounts: customerCounts.bySegment || {},
     count,
     kind: message ? message.kind : kind,
-    brands
+    brands,
+    message,
+    save,
+    users
   };
 
   return <AutoAndManualForm {...updatedProps} />;
 };
 
 AutoAndManualFormContainer.propTypes = {
+  messageId: PropTypes.string,
+  history: PropTypes.object,
   segmentsQuery: PropTypes.object,
   emailTemplatesQuery: PropTypes.object,
   customerCountsQuery: PropTypes.object,
@@ -80,25 +122,32 @@ AutoAndManualFormContainer.propTypes = {
   segmentsAddQuery: PropTypes.func,
   kind: PropTypes.string,
   engageMessageDetailQuery: PropTypes.object,
-  brandsQuery: PropTypes.object
+  brandsQuery: PropTypes.object,
+  usersQuery: PropTypes.object,
+  addMutation: PropTypes.func,
+  editMutation: PropTypes.func
 };
 
-export default withFormMutations(
-  compose(
-    graphql(gql(queries.emailTemplates), { name: 'emailTemplatesQuery' }),
-    graphql(gql(queries.segments), { name: 'segmentsQuery' }),
-    graphql(gql(queries.customerCounts), { name: 'customerCountsQuery' }),
-    graphql(gql(queries.headSegments), { name: 'headSegmentsQuery' }),
-    graphql(gql(queries.combinedFields), { name: 'combinedFieldsQuery' }),
-    graphql(gql(mutations.segmentsAdd), { name: 'segmentsAddQuery' }),
-    graphql(gql(queries.engageMessageDetail), {
-      name: 'engageMessageDetailQuery',
-      options: ({ messageId }) => ({
-        variables: {
-          _id: messageId
-        }
-      })
-    }),
-    graphql(gql(queries.brands), { name: 'brandsQuery' })
-  )(AutoAndManualFormContainer)
-);
+const EngageForm = compose(
+  graphql(gql(queries.users), { name: 'usersQuery' }),
+  graphql(gql(mutations.messagesAdd), { name: 'addMutation' }),
+  graphql(gql(mutations.messagesEdit), { name: 'editMutation' }),
+  graphql(gql(queries.emailTemplates), { name: 'emailTemplatesQuery' }),
+  graphql(gql(queries.segments), { name: 'segmentsQuery' }),
+  graphql(gql(queries.customerCounts), { name: 'customerCountsQuery' }),
+  graphql(gql(queries.headSegments), { name: 'headSegmentsQuery' }),
+  graphql(gql(queries.combinedFields), { name: 'combinedFieldsQuery' }),
+  graphql(gql(mutations.segmentsAdd), { name: 'segmentsAddQuery' }),
+  graphql(gql(queries.engageMessageDetail), {
+    name: 'engageMessageDetailQuery',
+    options: ({ messageId }) => ({
+      fetchPolicy: 'network-only',
+      variables: {
+        _id: messageId
+      }
+    })
+  }),
+  graphql(gql(queries.brands), { name: 'brandsQuery' })
+)(AutoAndManualFormContainer);
+
+export default withRouter(EngageForm);
